requirements.txt
ansible==2.9.9
ruamel.yaml
